+++ R/nn.R
@@ -419,3 +419,119 @@ predict.nn <- function(
 #' @export
 print.nn.predict <- function(x, ..., n = 10)
   print_predict_model(x, ..., n = n, header = "Neural Network")
+
+#' Cross-validation for Neural Network
+#'
+#' @details See \url{https://radiant-rstats.github.io/docs/model/nn.html} for an example in Radiant
+#'
+#' @param object Object of type "nn" or "nnet"
+#' @param K Number of cross validation passes to use
+#' @param repeats Repeated cross validation
+#' @param size Number of units (nodes) in the hidden layer
+#' @param decay Parameter decay
+#' @param seed Random seed to use as the starting point
+#' @param trace Print progress
+#' @param fun Function to use for model evaluation (i.e., auc for classification and RMSE for regression)
+#' @param ... Additional arguments to be passed to 'fun'
+#'
+#' @return A data.frame sorted by the mean of the performance metric
+#'
+#' @seealso \code{\link{summary.nn}} to summarize results
+#' @seealso \code{\link{plot.nn}} to plot results
+#' @seealso \code{\link{predict.nn}} for prediction
+#'
+#' @importFrom nnet nnet.formula
+#' @importFrom shiny getDefaultReactiveDomain withProgress incProgress
+#'
+#' @export
+cv.nn <- function(
+  object, K = 5, repeats = 1, decay = seq(0, 1, .2), size = 1:5,
+  seed = 1234, trace = TRUE, fun, ...
+) {
+
+  if (inherits(object, "nn")) object <- object$model
+  if (inherits(object, "nnet")) {
+    dv <- as.character(object$call$formula[[2]])
+    m <- eval(object$call[["data"]])
+    if (is.numeric(m[[dv]])) {
+      type <- "regression"
+    } else {
+      type <- "classification"
+      if (is.factor(m[[dv]])) {
+        lev <- levels(m[[dv]])[1]
+      } else if (is.logical(m[[dv]])) {
+        lev <- TRUE
+      } else {
+        stop("The level to use for classification is not clear. Use a factor of logical as the response variable")
+      }
+    }
+  } else {
+    stop("The model object does not seems to be a neural network")
+  }
+
+  set.seed(seed)
+  tune_grid <- expand.grid(decay = decay, size = size)
+  out <- data.frame(mean = NA, std = NA, min = NA, max = NA, decay = tune_grid[["decay"]], size = tune_grid[["size"]])
+
+  if (missing(fun)) {
+    if (type == "classification") {
+      fun = radiant.model::auc
+      cn <- "AUC (mean)"
+    } else {
+      fun = radiant.model::RMSE
+      cn <- "RMSE (mean)"
+    }
+  } else {
+    cn <- glue("{deparse(substitute(fun))} (mean)")
+  }
+
+  if (length(shiny::getDefaultReactiveDomain()) > 0) {
+    trace <- FALSE
+    incProgress <- shiny::incProgress
+    withProgress <- shiny::withProgress
+  } else {
+    incProgress <- function(...) {}
+    withProgress <- function(...) list(...)[["expr"]]
+  }
+
+  nitt <- nrow(tune_grid)
+  withProgress(message = "Running cross-validation (nn)", value = 0, {
+    for (i in seq_len(nitt)) {
+      perf <- double(K * repeats)
+      object$call[["decay"]] <- tune_grid[i, "decay"]
+      object$call[["size"]] <- tune_grid[i, "size"]
+      if (trace) cat("Working on size", tune_grid[i, "size"], "decay", tune_grid[i, "decay"], "\n")
+      for (j in seq_len(repeats)) {
+        rand <- sample(K, nrow(m), replace = TRUE)
+        for (k in seq_len(K)) {
+          object$call[["data"]] <- quote(m[rand != k, , drop = FALSE])
+          pred <- predict(eval(object$call), newdata = m[rand == k, , drop = FALSE])[,1]
+          if (type == "classification") {
+            if (missing(...)) {
+              perf[k + (j - 1) * K] <- fun(pred, unlist(m[rand == k, dv]), lev)
+            } else {
+              perf[k + (j - 1) * K] <- fun(pred, unlist(m[rand == k, dv]), lev, ...)
+            }
+          } else {
+            if (missing(...)) {
+              perf[k + (j - 1) * K] <- fun(pred, unlist(m[rand == k, dv]))
+            } else {
+              perf[k + (j - 1) * K] <- fun(pred, unlist(m[rand == k, dv], ...))
+            }
+          }
+        }
+      }
+      out[i,1:4] <- c(mean(perf), sd(perf), min(perf), max(perf))
+      incProgress(1/nitt, detail = paste("\nCompleted run", i, "out of", nitt))
+    }
+  })
+
+  if (type == "classification") {
+    out <- arrange(out, desc(mean))
+  } else {
+    out <- arrange(out, mean)
+  }
+  ## show evaluation metric in column name
+  colnames(out)[1] <- cn
+  out
+}